+++ BE/airbnb-be/src/main/java/com/codesquad/airbnb/accommodation/service/AccommodationService.java
@@ -2,29 +2,9 @@ package com.codesquad.airbnb.accommodation.service;
 
 import com.codesquad.airbnb.accommodation.controller.AccommodationRequestDTO;
 import com.codesquad.airbnb.accommodation.controller.AccommodationResponseDTO;
-import com.codesquad.airbnb.common.utils.DummyDataFactory;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-public class AccommodationService {
-    public List<AccommodationResponseDTO> readAll(AccommodationRequestDTO accommodationRequestDTO) {
-        List<AccommodationResponseDTO> result = DummyDataFactory.accommodationResponseDTOsWithId();
-
-        if (accommodationRequestDTO.getStartPrice() != null) {
-            result = result.stream()
-                             .filter(accommodationResponseDTO -> accommodationRequestDTO.getStartPrice() <= accommodationResponseDTO.pricePerNight())
-                             .collect(Collectors.toList());
-        }
-
-        if (accommodationRequestDTO.getEndPrice() != null) {
-            result = result.stream()
-                             .filter(accommodationResponseDTO -> accommodationResponseDTO.pricePerNight() <= accommodationRequestDTO.getEndPrice())
-                             .collect(Collectors.toList());
-        }
-
-        return result;
-    }
+public interface AccommodationService {
+    List<AccommodationResponseDTO> readAll(AccommodationRequestDTO accommodationRequestDTO);
 }
