+++ Procfile
@@ -1,1 +1,1 @@
-web: java -Dserver.port=$PORT $JAVA_OPTS -jar BE/airbnb-be/build/libs/*.jar
+web: java -Dspring.profiles.active=mockup -Dserver.port=$PORT $JAVA_OPTS -jar BE/airbnb-be/build/libs/*.jar
